Configuring how to pick trace. Closes #41
Added an enumerable configuration item, which either
can be set to RR_TRACE_DIR or "User provided"

When set to RR_TRACE_DIR, we use rr to list traces
otherwise the user is requested to input (full) path
to the trace dir when hitting "start".

For future features, maybe implementing a configuration
item for explicitly pointing to a path every time
is desirable.